backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
chromadb==0.6.3
langgraph==0.2.74
langchain-core==0.3.34
openai==1.61.1
httpx==0.28.1
python-dateutil==2.9.0.post0
pymysql==1.1.1
pytest==8.3.4
pytest-asyncio==0.25.3

